requirements.txt
python-dotenv
black
openai
anthropic
google.generativeai
streamlit
pandas
charset_normalizer
watchdog